FEATURE: Add color palette list selection to new UI
