+++ app/assets/javascripts/admin/addon/routes/admin-config-color-palettes.js
@@ -1,0 +1,8 @@
+import DiscourseRoute from "discourse/routes/discourse";
+import ColorScheme from "admin/models/color-scheme";
+
+export default class AdminConfigColorPalettesShowRoute extends DiscourseRoute {
+  model() {
+    return ColorScheme.findAll();
+  }
+}
