+++ app/assets/javascripts/admin/addon/templates/config-color-palettes-show.gjs
@@ -2,5 +2,5 @@ import RouteTemplate from "ember-route-template";
 import ColorPalette from "admin/components/admin-config-areas/color-palette";
 
 export default RouteTemplate(
-  <template><ColorPalette @colorPalette={{@controller.model}} /></template>
+  <template><ColorPalette @colorPalette={{@model}} /></template>
 );
