+++ app/assets/javascripts/admin/addon/templates/config-color-palettes.gjs
@@ -1,0 +1,6 @@
+import RouteTemplate from "ember-route-template";
+import ColorPalettes from "admin/components/admin-config-areas/color-palettes";
+
+export default RouteTemplate(
+  <template><ColorPalettes @palettes={{@model.content}} /></template>
+);
