+++ app/controllers/admin/config/color_palettes_controller.rb
@@ -1,5 +1,8 @@
 # frozen_string_literal: true
 class Admin::Config::ColorPalettesController < Admin::AdminController
+  def index
+  end
+
   def show
     render_serialized(ColorScheme.find(params[:id]), ColorSchemeSerializer, root: false)
   end
