+++ spec/system/admin_color_palettes_config_area_spec.rb
@@ -2,129 +2,36 @@
 
 describe "Admin Color Palettes Config Area Page", type: :system do
   fab!(:admin)
-  fab!(:color_scheme) { Fabricate(:color_scheme, user_selectable: false, name: "A Test Palette") }
+  fab!(:palette_1) { Fabricate(:color_scheme, user_selectable: false, name: "A Test Palette 1") }
+  fab!(:palette_2) { Fabricate(:color_scheme, user_selectable: false, name: "A Test Palette 2") }
 
   let(:config_area) { PageObjects::Pages::AdminColorPalettesConfigArea.new }
-  let(:toasts) { PageObjects::Components::Toasts.new }
+  let(:edit_config_area) { PageObjects::Pages::AdminColorPaletteConfigArea.new }
+  let(:create_color_palette_modal) { PageObjects::Modals::CreateColorPalette.new }
 
   before { sign_in(admin) }
 
-  it "allows editing the palette name" do
-    config_area.visit(color_scheme.id)
+  it "can navigate between different palettes" do
+    config_area.visit
 
-    config_area.edit_name_button.click
-    config_area.name_field.fill_in("Changed name 2.0")
+    config_area.palette(palette_1.id).click
+    expect(edit_config_area.palette_id).to eq(palette_1.id)
+    expect(page).to have_current_path("/admin/config/colors/#{palette_1.id}")
 
-    expect(config_area).to have_unsaved_changes_indicator
-
-    config_area.form.submit
-
-    expect(toasts).to have_success(I18n.t("js.saved"))
-    expect(config_area).to have_no_unsaved_changes_indicator
-    expect(color_scheme.reload.name).to eq("Changed name 2.0")
-    expect(config_area.name_heading.text).to eq("Changed name 2.0")
+    config_area.palette(palette_2.id).click
+    expect(edit_config_area.palette_id).to eq(palette_2.id)
+    expect(page).to have_current_path("/admin/config/colors/#{palette_2.id}")
   end
 
-  it "allows changing the user selectable field" do
-    config_area.visit(color_scheme.id)
+  it "can create new color palettes" do
+    config_area.visit
 
-    config_area.user_selectable_field.toggle
+    max_id = ColorScheme.maximum(:id) + 1
+    config_area.create_button.click
+    create_color_palette_modal.base_dropdown.select_row_by_name("Grey Amber")
+    create_color_palette_modal.create_button.click
 
-    expect(config_area).to have_unsaved_changes_indicator
-
-    config_area.form.submit
-
-    expect(toasts).to have_success(I18n.t("js.saved"))
-    expect(config_area).to have_no_unsaved_changes_indicator
-    expect(config_area.user_selectable_field.value).to eq(true)
-    expect(color_scheme.reload.user_selectable).to eq(true)
-  end
-
-  it "allows changing colors" do
-    config_area.visit(color_scheme.id)
-
-    expect(config_area.color_palette_editor).to have_light_tab_active
-
-    config_area.color_palette_editor.input_for_color("primary").fill_in(with: "#abcdef")
-
-    expect(config_area).to have_unsaved_changes_indicator
-
-    config_area.color_palette_editor.switch_to_dark_tab
-    expect(config_area.color_palette_editor).to have_dark_tab_active
-
-    config_area.color_palette_editor.input_for_color("primary").fill_in(with: "#fedcba")
-    config_area.color_palette_editor.input_for_color("secondary").fill_in(with: "#111222")
-
-    config_area.form.submit
-
-    expect(toasts).to have_success(I18n.t("js.saved"))
-    expect(config_area).to have_no_unsaved_changes_indicator
-    expect(config_area.color_palette_editor).to have_dark_tab_active
-    expect(config_area.color_palette_editor.input_for_color("primary").value).to eq("#fedcba")
-    expect(config_area.color_palette_editor.input_for_color("secondary").value).to eq("#111222")
-
-    config_area.color_palette_editor.switch_to_light_tab
-    expect(config_area.color_palette_editor).to have_light_tab_active
-    expect(config_area.color_palette_editor.input_for_color("primary").value).to eq("#abcdef")
-
-    expect(color_scheme.colors.find_by(name: "primary").hex).to eq("abcdef")
-    expect(color_scheme.colors.find_by(name: "primary").dark_hex).to eq("fedcba")
-    expect(color_scheme.colors.find_by(name: "secondary").dark_hex).to eq("111222")
-  end
-
-  it "allows duplicating the color palette" do
-    max_id = ColorScheme.maximum(:id)
-    color_scheme.update!(user_selectable: true)
-
-    config_area.visit(color_scheme.id)
-
-    expect(config_area.user_selectable_field.value).to eq(true)
-
-    config_area.duplicate_button.click
-
-    expect(page).to have_current_path("/admin/config/colors/#{max_id + 1}")
-    expect(config_area).to have_no_unsaved_changes_indicator
-    expect(config_area.name_heading.text).to eq(
-      I18n.t("admin_js.admin.config_areas.color_palettes.copy_of", name: color_scheme.name),
-    )
-    expect(toasts).to have_success(
-      I18n.t("admin_js.admin.config_areas.color_palettes.copy_created", name: color_scheme.name),
-    )
-    expect(config_area.user_selectable_field.value).to eq(false)
-  end
-
-  it "applies the changes live when editing the currently active palette" do
-    admin.user_option.update!(color_scheme_id: color_scheme.id)
-    config_area.visit(color_scheme.id)
-    config_area.color_palette_editor.input_for_color("secondary").fill_in(with: "#aa339f")
-
-    expect(config_area).to have_unsaved_changes_indicator
-    config_area.form.submit
-    expect(toasts).to have_success(I18n.t("js.saved"))
-    expect(config_area).to have_no_unsaved_changes_indicator
-
-    href = Stylesheet::Manager.new.color_scheme_stylesheet_link_tag_href(color_scheme.id)
-
-    expect(page).to have_css(
-      "link[data-scheme-id=\"#{color_scheme.id}\"][href=\"#{href}\"]",
-      visible: false,
-    )
-    expect(find("html").native.css_value("background-color")).to eq(
-      "rgba(#{"aa".to_i(16)}, #{"33".to_i(16)}, #{"9f".to_i(16)}, 1)",
-    )
-  end
-
-  it "doesn't apply changes when editing a palette that's not currently active" do
-    config_area.visit(color_scheme.id)
-    config_area.color_palette_editor.input_for_color("secondary").fill_in(with: "#aa339f")
-
-    expect(config_area).to have_unsaved_changes_indicator
-    config_area.form.submit
-    expect(toasts).to have_success(I18n.t("js.saved"))
-    expect(config_area).to have_no_unsaved_changes_indicator
-
-    href = Stylesheet::Manager.new.color_scheme_stylesheet_link_tag_href(color_scheme.id)
-
-    expect(page).to have_no_css("link[href=\"#{href}\"]", visible: false)
+    expect(page).to have_current_path("/admin/config/colors/#{max_id}")
+    expect(edit_config_area.palette_id).to eq(max_id)
   end
 end
